fix(viewport): remove the event binding only if viewportElement exists
The only case where viewportElement does not exist is in mult-workspace. In fact, this is because the Workbench component in designable/react uses workspace components directly. Workspace nesting occurs when users use multiple workspaces. It is recommended that direct use of Workspace be removed from Workbench components, or that Workbench components not be rendered as a separate UI